fix(): support keycloak-js v26.2.1
Storing the keycloak instance with `reactive()` and later converting it to a `ref()` using `toRefs()`
does not work with keycloak-js v26.2.1, because Keycloak methods (login, logout, register, etc.)
are not accessible in this way.
This commit stores the keycloak instance with `shallowRef()` which solves this issue and
improves performance, as the keycloak object does not need to be deeply reactive.